+++ latexonhttp/utils/misc.py
@@ -9,7 +9,7 @@
 """
 import subprocess
 
-CURRENT_API_VERSION = "2022-05-17-1"
+CURRENT_API_VERSION = "2022-05-17-3"
 
 
 def get_api_version():